Dictionary key JSON converters
