+++ src/Elastic.Clients.Elasticsearch/Common/Request/PlainRequestBase.cs
@@ -7,87 +7,6 @@ using System.Text.Json.Serialization;
 
 namespace Elastic.Clients.Elasticsearch
 {
-	//public readonly partial struct PropertyName : IDictionaryKey
-	//{
-	//	public string Key => Value;
-	//}
-
-	//// This is an incomplete stub implementation and should really be a struct
-	//public partial class Indices : IUrlParameter
-	//{
-	//	public static readonly Indices All = new("_all");
-
-	//	internal Indices(IndexName index) => _indexNameList.Add(index);
-
-	//	public Indices(IEnumerable<IndexName> indices)
-	//	{
-	//		indices.ThrowIfEmpty(nameof(indices));
-	//		_indexNameList.AddRange(indices);
-	//	}
-
-	//	public Indices(IEnumerable<string> indices)
-	//	{
-	//		indices.ThrowIfEmpty(nameof(indices));
-	//		_indexNameList.AddRange(indices.Select(s => (IndexName)s));
-	//	}
-
-	//	public IReadOnlyCollection<IndexName> Values => _indexNameList.ToArray();
-
-	//	public static Indices Parse(string names) => names.IsNullOrEmptyCommaSeparatedList(out var list) ? null : new Indices(list);
-
-	//	public static Indices Single(string index) => new Indices((IndexName)index);
-
-	//	public static implicit operator Indices(string names) => Parse(names);
-
-	//	string IUrlParameter.GetString(ITransportConfiguration? settings)
-	//	{
-	//		if (settings is not IElasticsearchClientSettings elasticsearchClientSettings)
-	//			throw new Exception(
-	//				"Tried to pass index names on query sting but it could not be resolved because no Elastic.Clients.Elasticsearch settings are available.");
-
-	//		var indices = _indexNameList.Select(i => i.GetString(settings)).Distinct();
-
-	//		return string.Join(",", indices);
-	//	}
-	//}
-
-	//public partial struct IndicesList : IUrlParameter
-	//{
-	//	//public static readonly IndicesList All = new("_all");
-
-	//	private readonly List<IndexName> _indices = new();
-
-	//	internal IndicesList(IndexName index) => _indices.Add(index);
-
-	//	public IndicesList(IEnumerable<IndexName> indices)
-	//	{
-	//		indices.ThrowIfEmpty(nameof(indices));
-
-	//		// De-duplicating during creation avoids cost when accessing the values.
-	//		foreach (var index in indices)
-	//			if (!_indices.Contains(index))
-	//				_indices.Add(index);
-	//	}
-
-	//	public IndicesList(string[] indices)
-	//	{
-	//		indices.ThrowIfEmpty(nameof(indices));
-
-	//		foreach (var index in indices)
-	//			if (!_indices.Contains(index))
-	//				_indices.Add(index);
-	//	}
-
-	//	public IReadOnlyCollection<IndexName> Values => _indices;
-
-	//	public static IndicesList Parse(string names) => names.IsNullOrEmptyCommaSeparatedList(out var list) ? null : new IndicesList(list);
-
-	//	public static implicit operator IndicesList(string names) => Parse(names);
-	//}
-
-	//public partial struct IndicesList { string IUrlParameter.GetString(ITransportConfiguration? settings) => ""; }
-
-
 	public abstract partial class PlainRequestBase<TParameters>
 	{
 		///<summary>Include the stack trace of returned errors.</summary>
