+++ src/Elastic.Clients.Elasticsearch/Serialization/DefaultRequestResponseSerializer.cs
@@ -55,7 +55,6 @@ internal class DefaultRequestResponseSerializer : SystemTextJsonSourceSerializer
 					new SelfDeserializableConverterFactory(settings),
 					new SelfTwoWaySerializableConverterFactory(settings),
 					new IndicesJsonConverter(settings),
-					new DictionaryConverter(settings),
 					new PropertyNameConverter(settings),
 					new IsADictionaryConverter(),
 					new ResponseItemConverterFactory(),
